+++ backend/src/config/@types/config.type.ts
