+++ backend/src/database/database.provider.ts
@@ -1,6 +1,6 @@
 import { Provider } from '@nestjs/common';
 import * as mongoose from 'mongoose';
-import { ConfigurationInterface } from '../config/config.interface';
+import { ConfigurationInterface } from '../config/@types/config.interface';
 import { DATABASE_CONNECTION } from './constants/database.constant';
 
 export const databaseProviders: Provider[] = [
